Added error catching, changed log base, etc.
Error catching accomplished with sub_mill(), to identify potential errors in dedupe_MA() when called within this function. Meta-analysis is now calculated with the natural logarithm of the odds ratio, rather than log base 10 of the odds ratio. When new meta-analysis rows are added, new values in the exclude column (for v1 only) are converted from logical to string characters, to adhere to the expected data type for that column & prevent error.