Added error catching, changed messages to the user
Error catching accomplished with sub_mill(), to identify potential errors in check_version(), get_col_types(), log_warnings(), handle_col_warnings(), and mill(), when called within this function. get_col_types(), log_warnings(), and handle_col_warnings() are used to check that input data meet the expected data types for each column, and, in the event that they do not, notify the user of which specific cells are causing such errors and give them the option to stop the pipeline and fix them. User messages were also changed to specify that the Enter key must be pressed to proceed past prompts that appear in the console.