Fix format_ma_results error
Call column must not be deleted from ma_results: otherwise the following error message occurs: "Names must be unique. These are duplicated: "b" at locations __ and __, ... beta, and se". This is a temporary fix, as the ma_results output (which only has the parameters used in meta-analysis, not the actual weighted average odds ratios for the models--those are in planks) still does not save to a csv, so this is not a complete fix. This change will allow planks to be saved, though.